Move App into own component directory with fixed imports
