+++ src/components/App/App.scss
@@ -1,4 +1,4 @@
-@import './partials/variables', './partials/extends';
+@import '../../partials/variables', '../../partials/extends';
 
 
 html {
