+++ src/components/SongCard/SongCard.scss
@@ -21,16 +21,6 @@
   }
 }
 
-// .song-card > img {
-//   grid-column: 1 / 2;
-//   border-radius: 50%;
-//   padding: 15px;
-//   width: 175px;
-//   height: 175px;
-//   justify-self: center;
-//   align-self: center;
-// }
-
 .song-details {
   text-align: left;
   grid-column: 2 / 3;
