+++ src/index.scss
@@ -4,7 +4,7 @@
 @import './partials/variables',
 './partials/extends',
 './partials/mixins',
-'./App.scss',
+'./components/App/App.scss',
 './components/SongCard/SongCard.scss',
 './components/Navigation/Navigation.scss',
 './components/SearchBar/SearchBar.scss',
