--- src/song-data.js
@@ -1,11 +1,0 @@
-const songData = [
-  {"id":1,"title":"Jigsaw falling into place","artist":"Radiohead","genres":["electronica","alternative rock"],"album_cover":"https://www.nme.com/wp-content/uploads/2016/09/62_radiohead-1.jpg"},
-  {"id":2,"title":"My Boo","artist":"Usher","genres":["Pop","R&B"],"album_cover":"https://upload.wikimedia.org/wikipedia/en/7/74/Usher_-_Confessions_album_cover.jpg"},
-  {"id":3,"title":"oops I did it again","artist":"Brittany Spears","genres":["Pop"],"album_cover":"https://images-na.ssl-images-amazon.com/images/I/81zgStH27mL._SL1500_.jpg"},
-  {"id":4,"title":"Body","artist":"Megan Thee Stallion","genres":["Rap","Hip Hop"],"album_cover":"https://media.pitchfork.com/photos/5fad9ae7cf46a2648df2494f/1:1/w_1000,h_1000,c_limit/Megan-Thee-Stallion.jpg"},
-  {"id":5,"title":"fly me to the moon","artist":"Frank Sinatra","genres":["Dance"],"album_cover":"https://static.qobuz.com/images/covers/9b/u4/qe9o9pdvuu49b_600.jpg"},
-  {"id":6,"title":"Bohemian Rhapsody","artist":"Queen","genres":["Rock","Metal"],"album_cover":"https://i.pinimg.com/originals/57/79/91/57799195bf2814769f27c598bb27e435.jpg"},
-  {"id": 7, "title":"I believe in a thing called love", "artist": "The Darkness", "genres": ["Rock"], "album_cover": "https://m.media-amazon.com/images/I/916MwTTbWOL._SS500_.jpg" }
-];
-
-export default songData;
